+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Fixed
+- lost loaded class statistic during import
+  ([issue #25](https://github.com/oboehm/ClazzFish/issues/25))
+
+
 ## [2.5] (22-Mar-2025)
 
 ### Added